Moved event file reading from files.py to eventfilereader.py Changed the structure of event file reading to have a parent class EventFileReader, from which child classes are produced for each type of EventFileReader e.g. HessioFileReader The correct file reader can be obtained from the input file (or specified by the user) by using EventFileReaderFactory Updated examples/display_integrator.py to use the EventFileReaderFactory Changed the default output directory path for examples/display_integrator.py to be inside the current directory
